# specs/001-task-crud-api/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Task CRUD API
  version: "1.0.0"
  description: >
    RESTful API for managing user-scoped tasks. User identity is accepted
    as a path parameter (user_id). No authentication enforcement in this
    feature scope.

paths:
  /api/users/{user_id}/todos:
    get:
      operationId: listTasks
      summary: List all tasks for a user
      description: Returns all tasks belonging to the specified user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
    post:
      operationId: createTask
      summary: Create a new task
      description: Creates a new task owned by the specified user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  /api/users/{user_id}/todos/{task_id}:
    get:
      operationId: getTask
      summary: Get a single task
      description: >
        Returns the task if it exists and belongs to the specified user.
        Returns 404 if the task does not exist or belongs to another user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTask
      summary: Partially update a task
      description: >
        Updates the specified fields of a task. Only fields included in the
        request body are modified. Returns 404 if the task does not exist
        or belongs to another user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
    delete:
      operationId: deleteTask
      summary: Delete a task
      description: >
        Permanently deletes the task. Returns 404 if the task does not exist
        or belongs to another user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Task deleted (no content)
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title (required)
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Task description (optional)

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: New task title
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: New task description
        is_completed:
          type: boolean
          description: Completion status

    TaskResponse:
      type: object
      required:
        - id
        - title
        - is_completed
        - created_at
        - updated_at
        - user_id
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        is_completed:
          type: boolean
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user_id:
          type: string

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string